refactor: moving loading.tsx into components' folder
