+++ src/components/loading.tsx
@@ -4,13 +4,13 @@ const Loading = () => {
   return (
     <Stack sx={{ background: 'white'}} position="absolute" zIndex={1000} height="100%" width="100%" direction="column">
       <Skeleton // Header
-        sx={{ height: '20%', mb: 3 }}
+        sx={{ height: '20%', mb: 1 }}
         width="100%"
         animation="wave"
         variant="rounded"
       />
       <Skeleton // Main
-        sx={{ height: '50%', mb: 3 }}
+        sx={{ height: '55%', mb: 1 }}
         width="100%"
         animation="wave"
         variant="rounded"
